Move the TTS stuff to the perception side
